Ver 0.a.x 20240904 	1. 얼굴 모자이크 설정이 꺼져 있을 때 영상 녹화를 진행하지 않던 버그 수정 	2. 0x82 0x49(비녹화 통화) 명령 적용 버전 	3.벨소리 동작 변경 2차 수정 	- 2차 벨 스트리밍 대기 동작에서 2차에서도 받지 않으면 종료 소리 발생 	- 2차 벨 스트리밍 대기 후 정상 종료되지 않는 버그 수정 	4. Amp 온 동작 변경 	- Amp 시작 시 Mute 동작 추가

+++ setup.sh
@@ -3,10 +3,12 @@ if [ -e "/dev/shm/system.bin" ]; then
 	echo "F/W System Change"
 	flashcp /dev/shm/system.bin /dev/mtd5
 	echo "Flashcp Finish!!"
-else
-	echo "F/W Card Change"
-	rm /tmp/mnt/sdcard/isc_bak
-	mv /tmp/mnt/sdcard/isc /tmp/mnt/sdcard/isc_bak
+fi
+
+if [ -e "/dev/shm/isc" ]; then
+	echo "F/W SD Card Change"
+	# rm /tmp/mnt/sdcard/isc_bak
+	# mv /tmp/mnt/sdcard/isc /tmp/mnt/sdcard/isc_bak
 	cp /dev/shm/isc /tmp/mnt/sdcard/isc
 	chmod 777 /tmp/mnt/sdcard/isc
 fi
@@ -20,20 +22,30 @@ fi
 
 # Executation Update
 if [ -e "/dev/shm/vexe.sh" ]; then
-	echo "vexe.sh chnge start!"
+	echo "vexe.sh change start!"
 	cp /dev/shm/vexe.sh /tmp/mnt/sdcard
 fi
 
 # Mute Update
 if [ -e "/dev/shm/mute_msg.sh" ]; then
-	echo "Mute chnge start!"
+	echo "Mute change start!"
 	cp /dev/shm/mute_msg.sh /tmp/mnt/sdcard
 fi
 
 if [ -d "/dev/shm/effects" ]; then
-	echo "Sound File Cnage!"
+	echo "Sound File Change!"
 	cp /dev/shm/effects/*.wav /tmp/mnt/sdcard/effects
 fi
 
+# if [ -e "/etc/vexe.sh" ]; then
+# 	echo "Trash Clear for Flash System!"
+# 	rm /tmp/mnt/sdcard/isc
+# 	rm /tmp/mnt/sdcard/vexe.sh
+# 	rm /tmp/mnt/sdcard/must_msg.sh
+# fi
+
 sync
+
+sleep 3
+
 sync
